10-feature 병합 (close #10)

+++ setup.php
@@ -32,27 +32,33 @@
                             Util::ShowError(400, "Invalid user name or password");
                             Util::CloseDocument();
                         } else {
-                            if(mdl_user::MakeAdmin($uname, $upass) != FALSE) {
-                                echo '<script>alert("설정 완료! 이제 setup.php를 삭제해주세요."); location.href="./";</script>';
+                            if((array_key_exists('upasschk', $_POST) ? $_POST['upasschk'] : NULL) === $upass){
+                                if(mdl_user::MakeAdmin($uname, $upass) != FALSE) {
+                                    echo '<script>alert("설정 완료! 이제 setup.php를 삭제해주세요."); location.href="./";</script>';
+                                } else {
+                                    Util::ShowError(500, "DB Error");
+                                    Util::CloseDocument();
+                                }
                             } else {
-                                Util::ShowError(500, "DB Error");
-                                Util::CloseDocument();
+                                 echo '<script>alert("비밀번호와 비밀번호 확인이 일치하지 않습니다."); location.href="./setup.php";</script>';
                             }
                         }
-                        
                     } else {
                         print 
 <<<SETUPHTML
                         <form action="./setup.php" method="post">
-            			<div class="login-box">
+            			<div class="setup-box">
             				<span colspan=2 class="login-box-title">Chintomi 관리자 설정</span>
             				<table class="login-box-input">
             					<tr>
             						<td><input name="uname" class="form-control form-control-sm" type="text" placeholder="관리자 ID" tabindex="1"></td>
-            						<td rowspan=2><button type="submit" class="btn btn-primary" tabindex="3">등록</button></td>
+            						<td rowspan=3><button type="submit" class="btn btn-primary" tabindex="3">등록</button></td>
             					</tr>
             					<tr>
             						<td><input name="upass" class="form-control form-control-sm" type="password" placeholder="관리자 PW" tabindex="2"><td>
+            					</tr>
+            					<tr>
+            						<td><input name="upasschk" class="form-control form-control-sm" type="password" placeholder="관리자 PW 확인" tabindex="3"><td>
             					</tr>
             				</table>
             			</div>
